커밋

+++ src/main/java/com/spring/instagram/create/PostService.java
@@ -1,5 +1,6 @@
 package com.spring.instagram.create;
 
+import com.spring.instagram.login.SessionCheck;
 import com.spring.instagram.models.Post;
 import com.spring.instagram.models.PostItemModel;
 //import com.spring.instagram.models.PostItemRepository;
@@ -56,7 +57,8 @@ public class PostService {
     }
 
     @PostMapping
-    public String createPost(@RequestPart("body") String body, @RequestPart("file") MultipartFile file){
+    @SessionCheck
+    public String createPost(String userName,@RequestPart("body") String body, @RequestPart("file") MultipartFile file){
         try{
 //            this.postRepository.save(post);
 //            this.postRepository.save(new Post(asdfasf));
@@ -97,13 +99,16 @@ public class PostService {
         return "update OK";
     }
     @DeleteMapping(path="{id}")
-    public String deletePost(@PathVariable Long id){
+    @SessionCheck
+    public String deletePost(String userName,@PathVariable Long id){
         try {
             Optional<Post> postOptional = this.postRepository.findById(id);
-            if (postOptional.isPresent()) {
+            if (postOptional.isPresent()
+                    && postOptional.get().getWriteBy().getEmail().equals(userName)) {
+
                 this.postRepository.delete(postOptional.get());
             } else {
-                return "delete Fail";
+                return "Default fail";
             }
         } catch(Exception err) {
             return "Fail";
